# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ExecFleet API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/employee/rating": {
      "post": {
        "tags": [
          "Employees",
          "Rating"
        ],
        "summary": "Create a rating for a ride",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ride_id": {
                    "type": "string"
                  },
                  "rating": {
                    "type": "number",
                    "minimum": 1,
                    "maximum": 5
                  },
                  "comment": {
                    "type": "string"
                  }
                },
                "required": [
                  "ride_id",
                  "rating"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rating created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "rating": {
                          "type": "number"
                        },
                        "created_by": {
                          "type": "string"
                        },
                        "driver_id": {
                          "type": "string"
                        },
                        "timestamp": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/employee/verify-otp": {
      "post": {
        "tags": [
          "Employees"
        ],
        "summary": "Verify OTP for employee",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "otp": {
                    "type": "string",
                    "minLength": 6,
                    "maxLength": 6
                  }
                },
                "required": [
                  "email",
                  "otp"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verified successfully with token"
          },
          "401": {
            "description": "Invalid or expired OTP"
          }
        }
      }
    },
...
    "/api/employee/ride": {
      "post": {
        "tags": [
          "Ride"
        ],
        "summary": "Create a ride entry",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ride": {
                    "type": "string",
                    "enum": [
                      "go_with_you",
                      "go_by_myself"
                    ]
                  },
                  "user_id": {
                    "type": "string"
                  },
                  "company_id": {
                    "type": "string"
                  }
                },
                "required": [
                  "ride",
                  "user_id",
                  "company_id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ride created"
          },
          "400": {
            "description": "Validation or database error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/driver/verify-otp": {
      "post": {
        "tags": [
          "Drivers"
        ],
        "summary": "Verify driver's OTP and return JWT token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string",
                    "example": "03123456789"
                  },
                  "otp": {
                    "type": "string",
                    "minLength": 6,
                    "maxLength": 6,
                    "example": "123456"
                  }
                },
                "required": [
                  "phone",
                  "otp"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "OTP verified successfully"
                    },
                    "token": {
                      "type": "string",
                      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or expired OTP"
          },
          "401": {
            "description": "Invalid phone number or OTP"
          }
        }
      }
    },
...
  },
...
}